product/serializer.py
from rest_framework import serializers
from .models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'product_title', 'product_description', 'product_price', 'trace_date', 'trace_count']
